Organiza CRUD de usuários.
Contribui para a issue thiagobitencourt/webdev-2#55
