+++ MatheusEduardo/models/usuario.js
@@ -1,0 +1,4 @@
+'use strict';
+
+// TODO: Implementar.
+const mongoose = require('mongoose');
